+++ LoliaFrpClient/Core/User/Traffic/Tunnels/TunnelsGetResponse_data_tunnels.cs
@@ -33,11 +33,11 @@ namespace LoliaFrpClient.Core.User.Traffic.Tunnels
         public string Remark { get; set; }
 #endif
         /// <summary>The total_in property</summary>
-        public int? TotalIn { get; set; }
+        public long? TotalIn { get; set; }
         /// <summary>The total_out property</summary>
-        public int? TotalOut { get; set; }
+        public long? TotalOut { get; set; }
         /// <summary>The total_traffic property</summary>
-        public int? TotalTraffic { get; set; }
+        public long? TotalTraffic { get; set; }
         /// <summary>The tunnel_name property</summary>
 #if NETSTANDARD2_1_OR_GREATER || NETCOREAPP3_1_OR_GREATER
 #nullable enable
@@ -74,9 +74,9 @@ namespace LoliaFrpClient.Core.User.Traffic.Tunnels
                 { "max_connections", n => { MaxConnections = n.GetIntValue(); } },
                 { "node_id", n => { NodeId = n.GetStringValue(); } },
                 { "remark", n => { Remark = n.GetStringValue(); } },
-                { "total_in", n => { TotalIn = n.GetIntValue(); } },
-                { "total_out", n => { TotalOut = n.GetIntValue(); } },
-                { "total_traffic", n => { TotalTraffic = n.GetIntValue(); } },
+                { "total_in", n => { TotalIn = n.GetLongValue(); } },
+                { "total_out", n => { TotalOut = n.GetLongValue(); } },
+                { "total_traffic", n => { TotalTraffic = n.GetLongValue(); } },
                 { "tunnel_name", n => { TunnelName = n.GetStringValue(); } },
             };
         }
@@ -90,9 +90,9 @@ namespace LoliaFrpClient.Core.User.Traffic.Tunnels
             writer.WriteIntValue("max_connections", MaxConnections);
             writer.WriteStringValue("node_id", NodeId);
             writer.WriteStringValue("remark", Remark);
-            writer.WriteIntValue("total_in", TotalIn);
-            writer.WriteIntValue("total_out", TotalOut);
-            writer.WriteIntValue("total_traffic", TotalTraffic);
+            writer.WriteLongValue("total_in", TotalIn);
+            writer.WriteLongValue("total_out", TotalOut);
+            writer.WriteLongValue("total_traffic", TotalTraffic);
             writer.WriteStringValue("tunnel_name", TunnelName);
             writer.WriteAdditionalData(AdditionalData);
         }
